refactor(type): Extract TypeKind and TypeTraits into their own header
Reaching TypeKind or TypeTraits currently means parsing all ~2400 lines
of Type.h, which brings folly/Random.h, folly/container/F14Set.h,
common/serialization/Serializable.h, glog, gflags, <iomanip> and the
virtual Type hierarchy with it. 232 files include Type.h directly, and
many of them want nothing more than the kind enum or a trait off it --
CppToType<T>::name, for instance, resolves to
TypeTraits<KIND>::name, a static constexpr const char*.

The block is cleanly separable. Lines 73-380 of Type.h are pure
constexpr/POD and depend on none of its 27 includes: the TypeKind enum,
the 18 TypeTraits specializations, UnknownValue, is_string_kind /
is_nested_kind, and forward declarations of the Type impl classes. They
move to the new velox/type/TypeKind.h, whose only dependencies are
<compare>, <cstdint>, <limits>, <memory>, <string> and
common/EnumDeclare.h, plus declarations of StringView and Timestamp --
TypeTraits names those two solely to form type aliases, so incomplete
types suffice.

Type.h includes the new header, so all existing includers are
unaffected and no call site changes. The duplicate
`using int128_t = __int128_t;` at Type.h:53 moves along with it;
HugeInt.h declares the same alias independently.

The overview comment above the extracted block moves with it. It
introduced the TypeKind enum that directly followed, so after the
extraction it described nothing, and its claim that "this file has
multiple C++ type definitions ... - TypeKind" was no longer true. The
TypeKind entry now points at its new home, and TypeKind.h carries a
header comment naming the same type set, so the two files reference each
other rather than each describing the split.

VELOX_FLUENT_CAST stays in Type.h: it is defined, used, and #undef'd
within the Type class it generates members for, and expands to
TypeTraits<KIND>::ImplType, which Type.h still sees through this header.

Pure extraction. No behavior change, no API change.

Verified: full velox build clean; TypeKind.h and Type.h each compile
standalone; velox_type_test 288, velox_functions_test 1926,
velox_functions_spark_test 707, velox_expression_test 1636 -- 4557
tests, all passing.
