+++ velox/type/CMakeLists.txt
@@ -57,6 +57,7 @@ velox_add_library(
   Tokenizer.h
   Tree.h
   Type.h
+  TypeKind.h
   TypeCoercer.h
   TypeEncodingUtil.h
   TypeSerde.h
