+++ velox/type/Type.h
@@ -41,6 +41,7 @@
 #include "velox/type/HugeInt.h"
 #include "velox/type/StringView.h"
 #include "velox/type/Timestamp.h"
+#include "velox/type/TypeKind.h"
 #include "velox/type/Tree.h"
 #include "velox/type/tz/TimeZoneMap.h"
 
@@ -49,8 +50,6 @@ struct dynamic;
 }
 
 namespace facebook::velox {
-
-using int128_t = __int128_t;
 
 using column_index_t = uint32_t;
 
@@ -61,323 +60,14 @@ constexpr column_index_t kConstantChannel =
 /// BOOLEAN, TINYINT, SMALLINT, INTEGER, BIGINT, HUGEINT, REAL, DOUBLE, VARCHAR,
 /// VARBINARY, TIMESTAMP, ARRAY, MAP, ROW
 ///
-/// This file has multiple C++ type definitions for each of these logical types.
+/// There are multiple C++ type definitions for each of these logical types.
 /// These logical definitions each serve slightly different purposes.
 /// These type sets are:
-/// - TypeKind
+/// - TypeKind, along with TypeTraits, in TypeKind.h
 /// - Type (RowType, BigIntType, ect.)
 /// - Templated Types (Row<T...>, Map<K, V>, ...)
 ///     C++ templated classes. Never instantiated, used to pass limited type
 ///     information into template parameters.
-
-/// Simple enum with type category.
-enum class TypeKind : int8_t {
-  BOOLEAN = 0,
-  TINYINT = 1,
-  SMALLINT = 2,
-  INTEGER = 3,
-  BIGINT = 4,
-  REAL = 5,
-  DOUBLE = 6,
-  VARCHAR = 7,
-  VARBINARY = 8,
-  TIMESTAMP = 9,
-  HUGEINT = 10,
-  // Enum values for ComplexTypes start after 30 to leave
-  // some values space to accommodate adding new scalar/native
-  // types above.
-  ARRAY = 30,
-  MAP = 31,
-  ROW = 32,
-  UNKNOWN = 33,
-  FUNCTION = 34,
-  OPAQUE = 35,
-  INVALID = 36
-};
-
-VELOX_DECLARE_ENUM_NAME(TypeKind);
-
-template <TypeKind KIND>
-class ScalarType;
-class ShortDecimalType;
-class LongDecimalType;
-class ArrayType;
-class MapType;
-class RowType;
-class FunctionType;
-class OpaqueType;
-class UnknownType;
-
-struct UnknownValue {
-  bool operator==(const UnknownValue& /* b */) const {
-    return true;
-  }
-
-  auto operator<=>(const UnknownValue& /* b */) const {
-    return std::strong_ordering::equal;
-  }
-};
-
-template <typename T>
-void toAppend(
-    const ::facebook::velox::UnknownValue& /* value */,
-    T* /* result */) {
-  // TODO Implement
-}
-
-template <TypeKind KIND>
-struct TypeTraits {};
-
-template <>
-struct TypeTraits<TypeKind::BOOLEAN> {
-  using ImplType = ScalarType<TypeKind::BOOLEAN>;
-  using NativeType = bool;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::BOOLEAN;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "BOOLEAN";
-};
-
-template <>
-struct TypeTraits<TypeKind::TINYINT> {
-  using ImplType = ScalarType<TypeKind::TINYINT>;
-  using NativeType = int8_t;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::TINYINT;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "TINYINT";
-};
-
-template <>
-struct TypeTraits<TypeKind::SMALLINT> {
-  using ImplType = ScalarType<TypeKind::SMALLINT>;
-  using NativeType = int16_t;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::SMALLINT;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "SMALLINT";
-};
-
-template <>
-struct TypeTraits<TypeKind::INTEGER> {
-  using ImplType = ScalarType<TypeKind::INTEGER>;
-  using NativeType = int32_t;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::INTEGER;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "INTEGER";
-};
-
-template <>
-struct TypeTraits<TypeKind::BIGINT> {
-  using ImplType = ScalarType<TypeKind::BIGINT>;
-  using NativeType = int64_t;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::BIGINT;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "BIGINT";
-};
-
-template <>
-struct TypeTraits<TypeKind::REAL> {
-  using ImplType = ScalarType<TypeKind::REAL>;
-  using NativeType = float;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::REAL;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "REAL";
-};
-
-template <>
-struct TypeTraits<TypeKind::DOUBLE> {
-  using ImplType = ScalarType<TypeKind::DOUBLE>;
-  using NativeType = double;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::DOUBLE;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "DOUBLE";
-};
-
-template <>
-struct TypeTraits<TypeKind::VARCHAR> {
-  using ImplType = ScalarType<TypeKind::VARCHAR>;
-  using NativeType = velox::StringView;
-  using DeepCopiedType = std::string;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::VARCHAR;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "VARCHAR";
-};
-
-template <>
-struct TypeTraits<TypeKind::TIMESTAMP> {
-  using ImplType = ScalarType<TypeKind::TIMESTAMP>;
-  using NativeType = Timestamp;
-  using DeepCopiedType = Timestamp;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::TIMESTAMP;
-  // isPrimitiveType in the type traits indicate whether it is a leaf type.
-  // So only types which have other sub types, should be set to false.
-  // Timestamp does not contain other types, so it is set to true.
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "TIMESTAMP";
-};
-
-template <>
-struct TypeTraits<TypeKind::HUGEINT> {
-  using ImplType = ScalarType<TypeKind::HUGEINT>;
-  using NativeType = int128_t;
-  using DeepCopiedType = NativeType;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::HUGEINT;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "HUGEINT";
-};
-
-template <>
-struct TypeTraits<TypeKind::VARBINARY> {
-  using ImplType = ScalarType<TypeKind::VARBINARY>;
-  using NativeType = velox::StringView;
-  using DeepCopiedType = std::string;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::VARBINARY;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "VARBINARY";
-};
-
-template <>
-struct TypeTraits<TypeKind::ARRAY> {
-  using ImplType = ArrayType;
-  using NativeType = void;
-  using DeepCopiedType = void;
-  static constexpr uint32_t minSubTypes = 1;
-  static constexpr uint32_t maxSubTypes = 1;
-  static constexpr TypeKind typeKind = TypeKind::ARRAY;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "ARRAY";
-};
-
-template <>
-struct TypeTraits<TypeKind::MAP> {
-  using ImplType = MapType;
-  using NativeType = void;
-  using DeepCopiedType = void;
-  static constexpr uint32_t minSubTypes = 2;
-  static constexpr uint32_t maxSubTypes = 2;
-  static constexpr TypeKind typeKind = TypeKind::MAP;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "MAP";
-};
-
-template <>
-struct TypeTraits<TypeKind::ROW> {
-  using ImplType = RowType;
-  using NativeType = void;
-  using DeepCopiedType = void;
-  static constexpr uint32_t minSubTypes = 1;
-  static constexpr uint32_t maxSubTypes = std::numeric_limits<char16_t>::max();
-  static constexpr TypeKind typeKind = TypeKind::ROW;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "ROW";
-};
-
-template <>
-struct TypeTraits<TypeKind::UNKNOWN> {
-  using ImplType = UnknownType;
-  using NativeType = UnknownValue;
-  using DeepCopiedType = UnknownValue;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::UNKNOWN;
-  static constexpr bool isPrimitiveType = true;
-  static constexpr bool isFixedWidth = true;
-  static constexpr const char* name = "UNKNOWN";
-};
-
-template <>
-struct TypeTraits<TypeKind::INVALID> {
-  using ImplType = void;
-  using NativeType = void;
-  using DeepCopiedType = void;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::INVALID;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "INVALID";
-};
-
-template <>
-struct TypeTraits<TypeKind::FUNCTION> {
-  using ImplType = FunctionType;
-  using NativeType = void;
-  using DeepCopiedType = void;
-  static constexpr uint32_t minSubTypes = 1;
-  static constexpr uint32_t maxSubTypes = std::numeric_limits<char16_t>::max();
-  static constexpr TypeKind typeKind = TypeKind::FUNCTION;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "FUNCTION";
-};
-
-template <>
-struct TypeTraits<TypeKind::OPAQUE> {
-  using ImplType = OpaqueType;
-  using NativeType = std::shared_ptr<void>;
-  using DeepCopiedType = std::shared_ptr<void>;
-  static constexpr uint32_t minSubTypes = 0;
-  static constexpr uint32_t maxSubTypes = 0;
-  static constexpr TypeKind typeKind = TypeKind::OPAQUE;
-  static constexpr bool isPrimitiveType = false;
-  static constexpr bool isFixedWidth = false;
-  static constexpr const char* name = "OPAQUE";
-};
-
-// Convenience constexpr function to check for string-like and nested type
-// kinds.
-constexpr bool is_string_kind(TypeKind kind) {
-  return kind == TypeKind::VARCHAR || kind == TypeKind::VARBINARY;
-}
-
-constexpr bool is_nested_kind(TypeKind kind) {
-  return kind == TypeKind::ARRAY || kind == TypeKind::MAP ||
-      kind == TypeKind::ROW;
-}
-
-template <TypeKind KIND>
-struct TypeFactory;
 
 #define VELOX_FLUENT_CAST(NAME, KIND)                                     \
   const typename TypeTraits<TypeKind::KIND>::ImplType& as##NAME() const { \
